requirements.txt
openpyxl==3.0.6
python-dotenv==0.15.0
requests==2.25.0
tqdm==4.54.1
tweepy==3.9.0
xlrd==1.2.0
